Move select all functions to global utilities
