+++ operators/utilities.py
@@ -1,5 +1,4 @@
 import logging
-from typing import Generator
 
 import bpy
 from bpy.types import Context, Operator
@@ -7,44 +6,11 @@ from mathutils import Matrix
 
 from .. import global_data
 from ..declarations import GizmoGroups, WorkSpaceTools
-from ..model.types import SlvsGenericEntity
 from ..convertors import update_convertor_geometry
 from ..utilities.preferences import use_experimental, get_prefs
+from ..utilities.data_handling import entities_3d
 
 logger = logging.getLogger(__name__)
-
-
-def entities_3d(context: Context) -> Generator[SlvsGenericEntity, None, None]:
-    for entity in context.scene.sketcher.entities.all:
-        if hasattr(entity, "sketch"):
-            continue
-        yield entity
-
-
-def select_all(context: Context):
-    sketch = context.scene.sketcher.active_sketch
-    if sketch:
-        logger.debug(
-            f"Selecting all sketcher entities in sketch : {sketch.name} (slvs_index: {sketch.slvs_index})"
-        )
-        generator = sketch.sketch_entities(context)
-    else:
-        logger.debug(f"Selecting all sketcher entities")
-        generator = entities_3d(context)
-
-    for e in generator:
-        if e.selected:
-            continue
-        if not e.is_visible(context):
-            continue
-        if not e.is_active(context.scene.sketcher.active_sketch):
-            continue
-        e.selected = True
-
-
-def deselect_all(context: Context):
-    logger.debug("Deselecting all sketcher entities")
-    global_data.selected.clear()
 
 
 def select_invert(context: Context):
